1.添加多租户处理

+++ src/Core/ViazyNetCore.Core/MultiTenancy/BasicTenantInfo.cs
@@ -9,14 +9,14 @@ namespace ViazyNetCore.MultiTenancy
 {
     public class BasicTenantInfo
     {
-        public long? TenantId { get; }
+        public int? TenantId { get; }
 
         /// <summary>
         /// Name of the tenant if <see cref="TenantId"/> is not null.
         /// </summary>
         public string Name { get; }
 
-        public BasicTenantInfo(long? tenantId, string name = null)
+        public BasicTenantInfo(int? tenantId, string name = null)
         {
             TenantId = tenantId;
             Name = name;
